requirements.txt
Pillow==9.5.0
streamlit==1.23.1